requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.9.2

# Configuration
django-environ==0.4.0
django-secure==1.0.1



# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0
django-floppyforms==1.6.1

# Models
django-model-utils==2.4

# Images
Pillow==3.1.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.24.1


# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.7

# Redis support
django-redis==4.3.0
redis>=2.10.0


celery==3.1.20


# Your custom requirements go here
django-datetime-widget==0.9.3
djangorestframework==3.3.2
